update
Renamed fast_filter to quick_filter, bug fix in fast_filter largest, test update, init changes for new names.
