+++ plantcv/plantcv/roi/__init__.py
@@ -9,9 +9,8 @@ from plantcv.plantcv.roi.roi_methods import custom
 from plantcv.plantcv.roi.roi_methods import filter
 from plantcv.plantcv.roi.roi_methods import auto_wells
 from plantcv.plantcv.roi.roi2mask import roi2mask
-from plantcv.plantcv.roi.quick_filter import quick_filter
-from plantcv.plantcv.roi.fast_filter import fast_filter, fast_rect_filter
+from plantcv.plantcv.roi.quick_filter import quick_filter, quick_rect_filter
 
 __all__ = ["circle", "ellipse", "from_binary_image", "rectangle", "auto_grid",
            "multi", "multi_rect", "custom", "auto_wells",
-           "filter", "roi2mask", "quick_filter", "fast_filter", "fast_rect_filter"]
+           "filter", "roi2mask", "quick_filter", "quick_filter", "quick_rect_filter"]
