+++ plantcv/plantcv/roi/quick_filter.py
@@ -1,17 +1,14 @@
-"""PlantCV quick_filter module."""
+"""PlantCV fast_filter module."""
 import os
 import cv2
 import numpy as np
-from skimage.measure import label
-from skimage.color import label2rgb
-from plantcv.plantcv._globals import params
-from plantcv.plantcv.roi import roi2mask
+from plantcv.plantcv import fatal_error
 from plantcv.plantcv._debug import _debug
-from plantcv.plantcv._helpers import _logical_operation
+from plantcv.plantcv._globals import params
 
 
 def quick_filter(mask, roi, roi_type="partial"):
-    """Quickly filter a binary mask using a region of interest.
+    """Filter a binary mask using a region of interest and connected components.
 
     Parameters
     ----------
@@ -20,111 +17,161 @@ def quick_filter(mask, roi, roi_type="partial"):
     roi : plantcv.plantcv.classes.Objects
         PlantCV ROI object.
     roi_type : str, optional
-        filter method, one of "partial", "cutto", or "within"
+        Type of ROI filtering: "partial", "cutto", "within", or "largest".
 
     Returns
     -------
     numpy.ndarray
         Filtered binary mask.
     """
-    # process for cutto is different, so if roi_type is cutto then use helper function
-    if roi_type == "cutto":
-        return _quick_cutto(mask, roi)[0]
-    # Increment the device counter
     params.device += 1
-
-    # Store debug
-    debug = params.debug
-    params.debug = None
-
-    # Label objects in the image from 1 to n (labeled mask)
-    labels, num = label(label_image=mask, return_num=True)
-
-    # Convert the input ROI to a binary mask (only works on single ROIs)
-    roi_mask = roi2mask(img=mask, roi=roi)
-
-    # Convert the labeled mask and ROI mask to float data types
-    roi_mask = roi_mask.astype(float)
-    labels = labels.astype(float)
-
-    # Set the ROI mask value to 0.5
-    roi_mask[np.where(roi_mask == 255)] = 0.5
-
-    # Add the labeled mask and ROI mask together
-    summed = roi_mask + labels
-
-    # For each label, if at least one pixel of the object overlaps the ROI
-    # set all the label values to the label plus 0.5
-    for i in range(1, num + 1):
-        # For each label, if at least one pixel of the object overlaps the ROI
-        # set all the label values to the label plus 0.5
-        if roi_type.upper() == "PARTIAL" and i + 0.5 in summed:
-            summed[np.where(summed == i)] = i + 0.5
-        # If one pixel of the object falls outside the ROI
-        # set all the label values to zero
-        elif roi_type.upper() == "WITHIN" and i in summed:
-            summed[np.where(labels == i)] = 0
-    # Objects that do not overlap the ROI will round to an integer and have
-    # the same value before and after rounding.
-    # Objects that overlap the ROI will round up/down and will not have the same value
-    # Where the values are equal (not overlapping)
-    summed[np.where(summed == summed.round())] = 0
-
-    # The summed image now only contains objects that overlap the ROI
-    # Subtract 0.5 to remove the ROI mask
-    summed = summed - 0.5
-
-    # Round and set the data type back to uint8
-    summed = summed.round().astype("uint8")
-
-    # Make sure the mask is binary
-    summed[np.where(summed > 0)] = 255
-
-    # Print/plot debug image
-    params.debug = debug
-    _debug(visual=summed, filename=os.path.join(params.debug_outdir, f"{params.device}_roi_filter.png"), cmap="gray")
-
-    return summed
+    roi_masks = _roi2masks(mask=mask, roi=roi)
+    filtered_mask = _filter_by_roi_masks(mask=mask, roi_masks=roi_masks, roi_type=roi_type)
+    _debug(visual=filtered_mask,
+           filename=os.path.join(params.debug_outdir, f"{params.device}_roi_filter.png"),
+           cmap="gray")
+    return filtered_mask
 
 
-def _quick_cutto(mask, roi):
-    """Quickly filter a binary mask using a region of interest by cutting to each ROI.
+def quick_rect_filter(mask, rois, roi_type="partial"):
+    """Filter a binary mask using one or more rectangular ROIs.
 
     Parameters
     ----------
     mask : numpy.ndarray
         Binary mask to filter.
-    roi : plantcv.plantcv.classes.Objects
-        PlantCV ROI object.
+    rois : list
+        Rectangular ROIs as `(x, y, width, height)` tuples.
+    roi_type : str, optional
+        Type of ROI filtering: "partial", "cutto", "within", or "largest".
 
     Returns
     -------
-    numpy.ndarray, numpy.ndarray, int
-        Filtered binary mask, labeled mask, number of labels.
+    numpy.ndarray
+        Filtered binary mask.
     """
-    # Increment the device counter
-    params.device += 1
+    roi_masks = _rects2masks(shape=mask.shape[:2], rois=rois)
+    filtered_mask = _filter_by_roi_masks(mask=mask, roi_masks=roi_masks, roi_type=roi_type)
+    _debug(visual=filtered_mask,
+           filename=os.path.join(params.debug_outdir, f"{params.device}_roi_filter.png"),
+           cmap="gray")
+    return filtered_mask
 
-    # Store debug
-    debug = params.debug
-    params.debug = None
 
-    mask_copy = np.copy(mask).astype(np.int32)
-    labeled_mask = np.zeros(mask.shape, dtype=np.int32)
-    bin_mask = np.copy(labeled_mask)
-    num_labels = len(roi.contours)
-    for i in range(num_labels):
-        # Pixel intensity of (i+1) such that the first object has value
-        cv2.drawContours(labeled_mask, roi.contours[i], -1, (i+1), -1)
-        cv2.drawContours(bin_mask, roi.contours[i], -1, (255), -1)
-    cropped_mask = _logical_operation(mask_copy, bin_mask, "and")
-    # Make a labeled mask from the cropped objects
-    label_mask_where = np.where(cropped_mask == 255, labeled_mask, 0)
+def _roi2masks(mask, roi):
+    roi_masks = []
+    for single_roi in roi:
+        roi_mask = np.zeros(mask.shape[:2], dtype=np.uint8)
+        cv2.drawContours(roi_mask, single_roi.contours[0], -1, 255, -1)
+        roi_masks.append(roi_mask)
+    return roi_masks
 
-    # Print/plot debug image
-    colorful = label2rgb(label_mask_where)
-    colorful2 = (255*colorful).astype(np.uint8)
-    params.debug = debug
-    _debug(visual=colorful2, filename=os.path.join(params.debug_outdir, f"{params.device}_label_colored_mask.png"))
 
-    return cropped_mask.astype(np.uint8), label_mask_where, num_labels
+def _rects2masks(shape, rois):
+    roi_masks = []
+    height, width = shape
+    for x, y, w, h in rois:
+        # clamp each rectangle corner to the image bounds so slicing below never goes out of range
+        x0 = max(0, min(width, int(round(x))))
+        y0 = max(0, min(height, int(round(y))))
+        x1 = max(0, min(width, int(round(x + w))))
+        y1 = max(0, min(height, int(round(y + h))))
+        if x1 > x0 and y1 > y0:
+            roi_mask = np.zeros(shape, dtype=np.uint8)
+            roi_mask[y0:y1, x0:x1] = 255
+            roi_masks.append(roi_mask)
+    return roi_masks
+
+
+def _filter_by_roi_masks(mask, roi_masks, roi_type):
+    roi_type = roi_type.lower()
+    binary = (mask > 0).astype(np.uint8) * 255
+    # elementwise max across all roi_masks, i.e. OR them together into one combined mask
+    roi_mask = np.maximum.reduce(roi_masks) if roi_masks else np.zeros(binary.shape[:2], dtype=np.uint8)
+
+    if roi_type == "cutto":
+        # keep only the binary-mask pixels that fall inside the ROI mask
+        return cv2.bitwise_and(binary, roi_mask)
+
+    if roi_type == "largest":
+        return _largest_in_each_roi(binary=binary, roi_masks=roi_masks)
+
+    # cv2.connectedComponentsWithStats: label each separate blob (8-connected neighbors) of the binary mask
+    # with a unique integer ID (0 = background); returns the count of labels, a label-ID image, per-label
+    # stats (bounding box/area), and centroids (_)
+    num_labels, labels, _stats, _ = cv2.connectedComponentsWithStats(binary, connectivity=8)
+    if num_labels <= 1:
+        return np.zeros(binary.shape, dtype=np.uint8)
+
+    inside = roi_mask > 0
+    # flatten the 2D labels image into a 1D array so np.bincount can
+    # tally values across every pixel
+    label_ids = labels.ravel()
+    # np.bincount: count, for each label ID, how many of its pixels are weighted by "inside";
+    # weights must be flattened (ravel) to line up with label ids.
+    overlap_counts = np.bincount(
+        label_ids,
+        weights=inside.ravel().astype(np.uint8),
+        minlength=num_labels,
+    )
+
+    if roi_type == "partial":
+        selected = overlap_counts > 0
+        selected[0] = False  # never select background label 0
+    elif roi_type == "within":
+        # count each component's pixels out of ROI
+        outside_counts = np.bincount(
+            label_ids,
+            weights=(~inside).ravel().astype(np.uint8),
+            minlength=num_labels,
+        )
+        # make an index of things that do overlap the ROI and do not have any pixels outside of the ROI
+        selected = (overlap_counts > 0) & (outside_counts == 0)
+        selected[0] = False  # never select background label 0
+    else:
+        fatal_error(f'ROI Type {roi_type} is not "cutto", "largest", "within" or "partial"!')
+
+    # make and return a binary mask of all the kept labels
+    return np.where(selected[labels], 255, 0).astype(np.uint8)
+
+
+def _largest_in_each_roi(binary, roi_masks):
+      output = np.zeros(binary.shape[:2], dtype=np.uint8)
+      # label the whole, uncut mask once — same as "partial" — so area and
+      # extent always refer to the full object, never a ROI-clipped sliver
+      num_labels, labels, stats, _ = cv2.connectedComponentsWithStats(binary, connectivity=8)
+      if num_labels <= 1:
+          return output
+      label_ids = labels.ravel()
+      for roi_mask in roi_masks:
+          inside = roi_mask > 0
+          # same overlap test "partial" uses: does this label touch the ROI at all?
+          overlap_counts = np.bincount(
+              label_ids,
+              weights=inside.ravel().astype(np.uint8),
+              minlength=num_labels,
+          )
+          candidates = np.flatnonzero(overlap_counts > 0)
+          candidates = candidates[candidates != 0]  # exclude background label 0
+          if candidates.size == 0:
+              continue
+          # rank candidates by their full area, not the portion inside the ROI
+          best = candidates[np.argmax(stats[candidates, cv2.CC_STAT_AREA])]
+          output[labels == best] = 255
+      return output
+
+
+def _legacy_largest_in_each_roi(binary, roi_masks):
+    output = np.zeros(binary.shape[:2], dtype=np.uint8)
+    for roi_mask in roi_masks:
+        # restrict the binary mask to just this one ROI's pixels
+        # Not how this should work, need to run partial THEN largest
+        clipped = cv2.bitwise_and(binary, roi_mask)
+        # label connected blobs within this ROI; stats gives area
+        num_labels, labels, stats, _ = cv2.connectedComponentsWithStats((clipped > 0).astype(np.uint8), connectivity=8)
+        if num_labels <= 1:
+            continue
+        # find the index of the roi with the largest area from cv2 above
+        best = 1 + int(np.argmax(stats[1:, cv2.CC_STAT_AREA]))
+        output[labels == best] = 255
+    return output
